// projects/stk/src/interfaces/api.ts
export interface IListServiceParams {
  token?: string
  locale?: string
  type?: 'template' | 'text' | 'svg' | 'background' | 'font' | 'marker' | 'layout' | 'theme' | 'hashtag' | 'group' | 'color'
  keyword?: string
  pageIndex?: number
  listAll?: 0 | 1
  listCategory?: 0 | 1
  listTag?: 0 | 1
  categoryIds?: number
  fontList?: 0 | 1 | 2
  aspect?: string
  width?: number
  height?: number
  theme?: string
  groupId?: string
  cache?: boolean
  platform?: string
  all_theme?: number
  colNum?: number,
  igLayout?: 'story' | 'post'
  shuffle?: number
}
// ...
export interface IListServiceContentData {
  id: number
  list: IListServiceContentDataItem[]
  title: string
  title_locale?: string // Category title locale
  is_recent?: number,
  cover_id?: string,
  cover_url?: string,
  url?: string
}

export interface IListServiceData {
  content: IListServiceContentData[]
  data: string
  host: string
  preview: string
  preview2?: string
  next_page?: number
  tags?: string[]
  /**
   * @param theme_ids - plz see https://www.notion.so/vivipic/Vivipic-35c05fc6c7e04d509ab7eb7a0f393fe4
   */
  theme_ids?: number[]
  url?: string // For IG tutorial viedo
}

export interface IListServiceResponse {
  data: IListServiceData
  flag: number
}